Fix: Changed Namespaces and moved project folders
