+++ src/BlazorElectron/BlazorElectron.sln
@@ -7,6 +7,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlazorElectron", "BlazorEle
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataSetlIOComponentTest", "..\ExcelIOComponentTest\DataSetlIOComponentTest.csproj", "{013AD4AA-D741-4ED8-8D72-B57A382ADB67}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PublicationDataSet", "..\PublicationDataSet\PublicationDataSet.csproj", "{22BA90BB-0173-41FF-BD5C-09A9DB30BDB0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExcelIO", "..\ExcelIO\ExcelIO.csproj", "{7B64B7FF-D94C-4D46-B956-DB9C616C0E32}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +25,14 @@ Global
 		{013AD4AA-D741-4ED8-8D72-B57A382ADB67}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{013AD4AA-D741-4ED8-8D72-B57A382ADB67}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{013AD4AA-D741-4ED8-8D72-B57A382ADB67}.Release|Any CPU.Build.0 = Release|Any CPU
+		{22BA90BB-0173-41FF-BD5C-09A9DB30BDB0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{22BA90BB-0173-41FF-BD5C-09A9DB30BDB0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{22BA90BB-0173-41FF-BD5C-09A9DB30BDB0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{22BA90BB-0173-41FF-BD5C-09A9DB30BDB0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7B64B7FF-D94C-4D46-B956-DB9C616C0E32}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7B64B7FF-D94C-4D46-B956-DB9C616C0E32}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7B64B7FF-D94C-4D46-B956-DB9C616C0E32}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7B64B7FF-D94C-4D46-B956-DB9C616C0E32}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
