--- src/BlazorElectron/Data/WeatherForecast.cs
@@ -1,15 +1,0 @@
-using System;
-
-namespace BlazorElectron.Data
-{
-    public class WeatherForecast
-    {
-        public DateTime Date { get; set; }
-
-        public int TemperatureC { get; set; }
-
-        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-
-        public string Summary { get; set; }
-    }
-}
