+++ src/BlazorElectron/Pages/AddPublication/AddAuthor.razor
@@ -1,5 +1,5 @@
 ﻿@page "/addAuthor"
-@using BlazorElectron.Data.PublicationDataSet
+@using Zeiss.Data.PublicationDataSetModel
 
 @inject NavigationManager NavigationManager
 @inject PublicationDataSetModel Publication
