+++ src/BlazorElectron/Pages/AddPublication/AddDateOfRelease.razor
@@ -1,5 +1,5 @@
 ﻿@page "/addDateOfRelease"
-@using BlazorElectron.Data.PublicationDataSet;
+@using Zeiss.Data.PublicationDataSetModel;
 @inject NavigationManager NavigationManager
 @inject PublicationDataSetModel Publication
