+++ src/BlazorElectron/Pages/AddPublication/AddDivision.razor
@@ -1,5 +1,5 @@
 ﻿@page "/addDivision"
-@using BlazorElectron.Data.PublicationDataSet;
+@using Zeiss.Data.PublicationDataSetModel;
 
 @inject NavigationManager NavigationManager
 @inject PublicationDataSetModel Publication
