+++ src/BlazorElectron/Pages/AddPublication/AddPublicationType.razor
@@ -1,6 +1,6 @@
 ﻿@page "/addPublicationType"
-@using BlazorElectron.Data.DataLogic;
-@using BlazorElectron.Data.PublicationDataSet;
+@using Zeiss.Data.DataLogic;
+@using Zeiss.Data.PublicationDataSetModel;
 
 @inject NavigationManager NavigationManager
 @inject PublicationDataSetModel Publication
