+++ src/BlazorElectron/Pages/AddPublication/AddPublisher.razor
@@ -1,6 +1,6 @@
 ﻿@page "/addPublisher"
-@using BlazorElectron.Data.DataLogic;
-@using BlazorElectron.Data.PublicationDataSet;
+@using Zeiss.Data.DataLogic;
+@using Zeiss.Data.PublicationDataSetModel;
 
 @inject NavigationManager NavigationManager
 @inject PublicationDataSetModel Publication
