+++ src/BlazorElectron/Pages/AddPublication/AddTitle.razor
@@ -1,5 +1,6 @@
 ﻿@page "/addTitle"
-@using BlazorElectron.Data.PublicationDataSet
+@using Zeiss.Data.PublicationDataSetModel
+
 
 @inject NavigationManager NavigationManager
 @inject PublicationDataSetModel Publication
