+++ src/BlazorElectron/Pages/LandingPage.razor
@@ -1,5 +1,5 @@
 ﻿@page "/"
-@using BlazorElectron.Data.PublicationDataSet;
+@using Zeiss.Data.PublicationDataSetModel;
 @inject NavigationManager NavigationManager
 @inject PublicationDataSetModel Publication
