+++ src/ExcelIO/ExcelIO.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
@@ -6,12 +6,9 @@
 
   <ItemGroup>
     <PackageReference Include="DocumentFormat.OpenXml" Version="2.12.3" />
-    <PackageReference Include="ElectronNET.API" Version="9.31.2" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\ExcelIOComponentTest\DataSetlIOComponentTest.csproj" />
-    <ProjectReference Include="..\ExcelIO\ExcelIO.csproj" />
     <ProjectReference Include="..\PublicationDataSet\PublicationDataSet.csproj" />
   </ItemGroup>
