--- src/ExcelIOComponentTest/Data/PublicationDataSet/Interface/IDataObject.cs
@@ -1,7 +1,0 @@
-﻿namespace Data.DataSet
-{
-    public interface IDataObject
-    {
-        int? ID { get; set; }
-    }
-}
