+++ src/PublicationDataSet/PublicationDataSet.csproj
