+++ src/PublicationDataSet/PublicationDataSet/PublicationDataSet.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Data.DataSet
+namespace Zeiss.Data.DataSet
 {
     public class PublicationDataSet : DataObject, IPublicationDataSet
     {
